Update URLEncodeing and Unit Test
